finished url sharing
